bootstrap: Add a CommonJS-style imports system
Use a native JS hook to add an implementation of the CommonJS Modules
1.1.1 specification, using require(). This is independent from the
existing imports system.

This implementation has one minor deviation from the spec to be more
compatible with the old imports system, and that is that modules are
not run in their own private namespace, but instead directly in their
"exports" namespace. That is, "this" and "exports" are the same object
within the module. The "module" variable is also on "exports".
This means that we can port the majority of modules over without
requiring that they be rewritten to use exports, which doesn't exist
in old versions of gjs.

The existing imports system is considered deprecated, and will soon
be rewritten as a compatibility layer in JS.